Rename estimation module
